[refactor] Pub/Sub 구조 추가 및 DockerEventCollector(Pub) - AutoHealingService(Sub) 분리, DockerEventCollector를 Repository로 분리하여 가독성 개선
